changing the project part-3
