+++ Service/src/main/java/TIcket/isPromo.java
@@ -2,5 +2,6 @@ package TIcket;
 
 public enum isPromo {
     YES,
-    NO
+    NO,
+    NOINFO;
 }
