+++ Service/src/main/java/org/example/classContent.java
@@ -2,7 +2,7 @@ package org.example;
 
 import java.io.FileWriter;
 
-public interface ClassInfo {
+public interface classContent {
     void print();
     void print(FileWriter file);
